requirements.txt
streamlit
google-generativai
python-dotenv
pdf2image
